Bug 4160: Currency conversion doesn't handle rates other than 100
Changes data type for currency.rate to accomodate bigger currency conversion rates.

FLOAT( 15, 5 )

+++ installer/data/mysql/kohastructure.sql
@@ -563,7 +563,7 @@ CREATE TABLE `currency` (
   `currency` varchar(10) NOT NULL default '',
   `symbol` varchar(5) default NULL,
   `timestamp` timestamp NOT NULL default CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
-  `rate` float(7,5) default NULL,
+  `rate` float(15,5) default NULL,
   `active` tinyint(1) default NULL,
   PRIMARY KEY  (`currency`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
